Split normalize code into two separate submodules
